auk filter - auk zf update2
We have updated the auk_filter step script from the filter request (e.g. bbox- sp ) to the eBird zf table. The inputs and outputs were adjusted and tested in ByB with the documentation revised by Lina and Helena. The next step is to integrate the pipeline with the covars script step. However The inputs, ebd_file=ebd_CO_relDec-2022 and ebd_sampling_file=ebd_samp_Col_relDec-2022-002, should be manually loaded into the repository/00_auk_filter/input directory, as their size is too large to be accommodated in GIT. We could consider downsizing the dataset to a specific region in order to provide a functional pilot without going through that step.